test: isolate converter timeout coverage
